Send POST request to the server
The code has been combined and tested and we made sure it POSTs onto the database, so it works! :)
Use these classes to actually post your registration to the server. 
Yoseph has written a class to send requests to the server as well which is integrated with the existing repo., hence we will be posting his code on to the master branch. 
Team A can use this as well if they want! All you will have to do is change the field Id's I have used to the one you guys are using in in the Main GUI for the registration page.