Login java (#84)
* updated check login details function
 done by vluo
pulls the JSON response received and checks login details when user attempts to login. If user exists pull the id and password and store it into db table. 

* Add files via upload

Class is only meant to send post request with login info to the login.php script
will be uploading hashing password and get requests soon so hold off on using this until I do that

*Merge branch 'login-Achintya' into login_java

* Login 
done by Achintya
This class will deal with sending post requests to the server and receiving the response.
Hashing yet to come :( sadly I am having difficulty with that. Also Vluo did a good job with receiving the responses.

* Add files via upload

* signinbugfix

couple of Bug fixes

* fixed merge conflicts with ash's branch