Signin and registration sync.
Today I noticed that Team A and Team B were working on a little bit of a different platform on the GUI part so I checked and fixed a little bit of the registration page and signin page to make sure the server is getting all the required login info.
I also have another signin page I was working on which works but for some reason does not mix well with the whole code so if you want to check that out please do. 
This also has brians GUI integrated with it. :)